chore: use execa replace node-pty
