+++ packages/iceworks-client/src/components/XtermTerminal/index.module.scss
@@ -1,0 +1,4 @@
+.xtermContainer {
+  width: 100%;
+  height: 100%;
+}
