+++ packages/iceworks-server/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "egg-scripts": "^2.10.0",
     "egg-socket.io": "^4.1.5",
+    "execa": "^1.0.0",
     "midway": "^1.0.0",
     "node-pty": "^0.8.1",
     "pty.js": "^0.3.1",
