+++ packages/iceworks-server/src/app/router.ts
@@ -1,4 +1,3 @@
 export default (app) => {
-  app.io.of('/').route('work', app.io.controller.work.index);
-  app.io.of('/').route('dev', app.io.controller.work.dev);
+  app.io.of('/').route('dev', app.io.controller.tasks.dev);
 };
